Move MenuGroup class to seperate file. Remove some instance variables from QtGame
